# Swagger.json
{
    "openapi": "3.1.0",
    "info": {
      "title": "Sistema de Subastas API",
      "description": "API REST del Sistema de Subastas — UADE Desarrollo de Aplicaciones I, 1C 2026. Gestiona registración de postores, catálogo, subastas, pujas, ventas, multas, seguros y solicitudes de incorporación de bienes.",
      "contact": {
        "name": "Equipo DA1",
        "email": "agustin.torres@majily.com"
      },
      "version": "0.1.0"
    },
    "paths": {
...
      "/rematadores": {
        "get": {
          "tags": [
            "Rematadores"
          ],
          "summary": "Listar Rematadores",
          "operationId": "listar_rematadores_rematadores_get",
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "items": {
                      "$ref": "#/components/schemas/RematadorOut"
                    },
                    "type": "array",
                    "title": "Response Listar Rematadores Rematadores Get"
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "Rematadores"
          ],
          "summary": "Crear Rematador",
          "operationId": "crear_rematador_rematadores_post",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RematadorCreate"
                }
              }
            },
            "required": true
          },
          "responses": {
            "201": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/RematadorOut"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/rematadores/{rematador_id}": {
        "get": {
          "tags": [
            "Rematadores"
          ],
          "summary": "Obtener Rematador",
          "operationId": "obtener_rematador_rematadores__rematador_id__get",
          "parameters": [
            {
              "name": "rematador_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Rematador Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/RematadorOut"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
...
      "/pujas": {
        "post": {
          "tags": [
            "Pujas"
          ],
          "summary": "Registrar una puja (valida reglas de incremento)",
          "operationId": "crear_puja_pujas_post",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PujaCreate"
                }
              }
            },
            "required": true
          },
          "responses": {
            "201": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/PujaOut"
                  }
                }
              }
            },
            "400": {
              "description": "Puja inválida (rango, usuario bloqueado, subasta cerrada, etc.)"
            },
            "404": {
              "description": "Ítem, subasta o usuario no encontrado"
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/pujas/item/{catalogo_item_id}": {
        "get": {
          "tags": [
            "Pujas"
          ],
          "summary": "Historial de pujas de un ítem (orden cronológico)",
          "operationId": "historial_item_pujas_item__catalogo_item_id__get",
          "parameters": [
            {
              "name": "catalogo_item_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Catalogo Item Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/PujaOut"
                    },
                    "title": "Response Historial Item Pujas Item  Catalogo Item Id  Get"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/pujas/item/{catalogo_item_id}/mejor": {
        "get": {
          "tags": [
            "Pujas"
          ],
          "summary": "Mejor oferta actual y rango válido para la próxima puja",
          "operationId": "mejor_oferta_pujas_item__catalogo_item_id__mejor_get",
          "parameters": [
            {
              "name": "catalogo_item_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Catalogo Item Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/MejorOferta"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/pujas/usuario/{usuario_id}": {
        "get": {
          "tags": [
            "Pujas"
          ],
          "summary": "Historial de pujas de un usuario",
          "operationId": "historial_usuario_pujas_usuario__usuario_id__get",
          "parameters": [
            {
              "name": "usuario_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Usuario Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/PujaOut"
                    },
                    "title": "Response Historial Usuario Pujas Usuario  Usuario Id  Get"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
...
      "/multas": {
        "post": {
          "tags": [
            "Multas"
          ],
          "summary": "Crear Multa",
          "operationId": "crear_multa_multas_post",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MultaCreate"
                }
              }
            },
            "required": true
          },
          "responses": {
            "201": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/MultaOut"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/multas/usuario/{usuario_id}": {
        "get": {
          "tags": [
            "Multas"
          ],
          "summary": "Listar Por Usuario",
          "operationId": "listar_por_usuario_multas_usuario__usuario_id__get",
          "parameters": [
            {
              "name": "usuario_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Usuario Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/MultaOut"
                    },
                    "title": "Response Listar Por Usuario Multas Usuario  Usuario Id  Get"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/multas/{multa_id}/pagar": {
        "post": {
          "tags": [
            "Multas"
          ],
          "summary": "Marcar multa como pagada",
          "operationId": "pagar_multa_multas__multa_id__pagar_post",
          "parameters": [
            {
              "name": "multa_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Multa Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/MultaOut"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "EstadoPuja": {
          "type": "string",
          "enum": [
            "pendiente",
            "confirmada",
            "rechazada"
          ],
          "title": "EstadoPuja"
        },
...
        "HTTPValidationError": {
          "properties": {
            "detail": {
              "items": {
                "$ref": "#/components/schemas/ValidationError"
              },
              "type": "array",
              "title": "Detail"
            }
          },
          "type": "object",
          "title": "HTTPValidationError"
        },
...
        "MejorOferta": {
          "properties": {
            "catalogo_item_id": {
              "type": "integer",
              "title": "Catalogo Item Id"
            },
            "mejor_monto": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mejor Monto"
            },
            "usuario_id": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Usuario Id"
            },
            "minimo_proxima": {
              "type": "string",
              "title": "Minimo Proxima"
            },
            "maximo_proxima": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Maximo Proxima"
            }
          },
          "type": "object",
          "required": [
            "catalogo_item_id",
            "minimo_proxima"
          ],
          "title": "MejorOferta"
        },
...
        "MultaCreate": {
          "properties": {
            "usuario_id": {
              "type": "integer",
              "title": "Usuario Id"
            },
            "venta_id": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Venta Id"
            },
            "monto": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "string"
                }
              ],
              "title": "Monto"
            },
            "moneda": {
              "type": "string",
              "title": "Moneda",
              "default": "ARS"
            },
            "motivo": {
              "type": "string",
              "title": "Motivo"
            }
          },
          "type": "object",
          "required": [
            "usuario_id",
            "monto",
            "motivo"
          ],
          "title": "MultaCreate"
        },
        "MultaOut": {
          "properties": {
            "id": {
              "type": "integer",
              "title": "Id"
            },
            "usuario_id": {
              "type": "integer",
              "title": "Usuario Id"
            },
            "venta_id": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Venta Id"
            },
            "monto": {
              "type": "string",
              "title": "Monto"
            },
            "moneda": {
              "type": "string",
              "title": "Moneda"
            },
            "motivo": {
              "type": "string",
              "title": "Motivo"
            },
            "pagada": {
              "type": "boolean",
              "title": "Pagada"
            },
            "fecha": {
              "type": "string",
              "format": "date-time",
              "title": "Fecha"
            }
          },
          "type": "object",
          "required": [
            "id",
            "usuario_id",
            "venta_id",
            "monto",
            "moneda",
            "motivo",
            "pagada",
            "fecha"
          ],
          "title": "MultaOut"
        },
        "PujaCreate": {
          "properties": {
            "catalogo_item_id": {
              "type": "integer",
              "title": "Catalogo Item Id"
            },
            "usuario_id": {
              "type": "integer",
              "title": "Usuario Id"
            },
            "monto": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "string"
                }
              ],
              "title": "Monto"
            }
          },
          "type": "object",
          "required": [
            "catalogo_item_id",
            "usuario_id",
            "monto"
          ],
          "title": "PujaCreate"
        },
        "PujaOut": {
          "properties": {
            "id": {
              "type": "integer",
              "title": "Id"
            },
            "subasta_id": {
              "type": "integer",
              "title": "Subasta Id"
            },
            "catalogo_item_id": {
              "type": "integer",
              "title": "Catalogo Item Id"
            },
            "usuario_id": {
              "type": "integer",
              "title": "Usuario Id"
            },
            "monto": {
              "type": "string",
              "title": "Monto"
            },
            "fecha_hora": {
              "type": "string",
              "format": "date-time",
              "title": "Fecha Hora"
            },
            "estado": {
              "$ref": "#/components/schemas/EstadoPuja"
            }
          },
          "type": "object",
          "required": [
            "id",
            "subasta_id",
            "catalogo_item_id",
            "usuario_id",
            "monto",
            "fecha_hora",
            "estado"
          ],
          "title": "PujaOut"
        },
        "RematadorCreate": {
          "properties": {
            "nombre": {
              "type": "string",
              "title": "Nombre"
            },
            "apellido": {
              "type": "string",
              "title": "Apellido"
            },
            "matricula": {
              "type": "string",
              "title": "Matricula"
            }
          },
          "type": "object",
          "required": [
            "nombre",
            "apellido",
            "matricula"
          ],
          "title": "RematadorCreate"
        },
        "RematadorOut": {
          "properties": {
            "id": {
              "type": "integer",
              "title": "Id"
            },
            "nombre": {
              "type": "string",
              "title": "Nombre"
            },
            "apellido": {
              "type": "string",
              "title": "Apellido"
            },
            "matricula": {
              "type": "string",
              "title": "Matricula"
            }
          },
          "type": "object",
          "required": [
            "id",
            "nombre",
            "apellido",
            "matricula"
          ],
          "title": "RematadorOut"
        },
...
        "ValidationError": {
          "properties": {
            "loc": {
              "items": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "integer"
                  }
                ]
              },
              "type": "array",
              "title": "Location"
            },
            "msg": {
              "type": "string",
              "title": "Message"
            },
            "type": {
              "type": "string",
              "title": "Error Type"
            }
          },
          "type": "object",
          "required": [
            "loc",
            "msg",
            "type"
          ],
          "title": "ValidationError"
        },
...
      }
    }
  }